requirements/development.txt
-r base.txt

# Development tools
django-debug-toolbar>=6.0.0
django-extensions>=4.1.0
ipython>=9.5.0

# Testing
pytest>=8.4.2
pytest-django>=4.11.1
pytest-cov>=7.0.0
factory-boy>=3.3.3
faker>=37.6.0

# Code Quality
ruff>=0.12.12
mypy>=1.17.1
pre-commit>=4.3.0
bandit>=1.8.6
black>=25.1.0
isort>=6.0.1
safety>=3.6.1

# Type stubs
django-stubs>=5.2.2
djangorestframework-stubs>=3.16.2
types-requests>=2.32.4

# Documentation
Sphinx>=8.2.3
sphinx-rtd-theme>=3.0.2
